requirements.txt
cffi==1.15.1
click==8.1.3
colorama==0.4.5
colorzero==2.0
cryptography==38.0.1
et-xmlfile==1.1.0
Flask==2.2.2
Flask-WTF==1.0.1
gpiozero==1.6.2
itsdangerous==2.1.2
Jinja2==3.1.2
jwt==1.3.1
MarkupSafe==2.1.1
openpyxl==3.0.10
pycparser==2.21
PyJWT==2.4.0
python-dotenv==0.21.0
Werkzeug==2.2.2
WTForms==3.0.1
